src/requirements.txt
#conda create -n llmbook python==3.9
#pip install torch==2.3.1 torchvision==0.18.1 torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu121
#pip install transformers[torch]==4.44.1
# git clone https://github.com/PanQiWei/AutoGPTQ.git && cd AutoGPTQ
# pip install -vvv --no-build-isolation -e .
# pip install -U sentence-transformers or pip install sentence-transformers[train]==3.0.1
#pip install -r src/requirements.txt --use-deprecated=legacy-resolver
# watch -d -n 0.5 nvidia-smi

torch==2.3.1+cu121
torchvision==0.18.1+cu121
torchaudio==2.3.1+cu121
transformers==4.44.1
datasets==2.19.1
jupyter==1.0.0
python-dotenv==1.0.1
accelerate==0.33.0
bitsandbytes==0.43.3
tiktoken==0.6.0
autotrain-advanced==0.8.13
peft==0.12.0
auto-gptq==0.8.0.dev0+cu121 #0.7.1
autoawq==0.2.5
huggingface_hub==0.23.4
optimum==1.17.1
aiohttp==3.10.5
vllm==0.5.3.post1 #0.5.2
llama-index==0.10.40 #0.10.34  #0.11.2 #
langchain-openai==0.1.6
nemoguardrails[openai]==0.8.0
openai==1.42.0 #1.25.1
chromadb==0.5.0
wandb==0.16.6
llama-index-callbacks-wandb==0.1.2
faiss-cpu==1.8.0
llama-index-embeddings-huggingface==0.2.3  # >=0.2.0  #==0.2.0 # 0.3.1
sentence-transformers==3.0.1